CountFactors task

+++ codility-solutions/src/main/java/ru/roman/task/codility/a010_prime_and_composite/Solution1CountFactors.java
@@ -26,8 +26,8 @@ class Solution1CountFactors {
         }
 
         int count = 2;
-        int last = -1;
-        for (int i = 2; i * i <= N; i++) {
+        long last = -1;
+        for (long i = 2; i * i <= N; i++) {
             if (N % i == 0) {
                 count += 2;
                 last = i;
